+++ src/ml_tooling/data/base_data.py
@@ -22,7 +22,10 @@ class Dataset(metaclass=abc.ABCMeta):
     train_y: Optional[DataType] = None
     train_x: Optional[pd.DataFrame] = None
     cached_data: Optional[pd.DataFrame] = None
-    plot: DataVisualize = None
+
+    @property
+    def plot(self):
+        return DataVisualize(self)
 
     def create_train_test(
         self,
@@ -70,7 +73,6 @@ class Dataset(metaclass=abc.ABCMeta):
     def x(self):
         if self._x is None:
             self._x, self._y = indexable(*self._load_training_data())
-            self.plot = DataVisualize(self)
         return self._x
 
     @x.setter
@@ -81,7 +83,6 @@ class Dataset(metaclass=abc.ABCMeta):
     def y(self):
         if self._y is None:
             self._x, self._y = indexable(*self._load_training_data())
-            self.plot = DataVisualize(self)
         return self._y
 
     @y.setter