+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - The dataframe returned by `.make_prediction`now labels the columns in a more
 human friendly manner
 - Dataset now verifies that `load_training_data` and `load_prediction_data` do not return empty
+- Added a missing data visualization to `Dataset.plot`
 
 # v0.10.2
 - Fixed type inferences from data to sql in _load_data